Add day08
Small refactor. This solution works only beacuse of the input. The length from 'A' to first 'Z' is the same as the cycle from the first 'Z' back to itself. So the solution is the LCM of the cycles. If there are intermidiate 'Z' inside the loop and the distance from 'A' to 'Z' is diffrent then the cycle 'Z' to 'Z' this will not work